refactor: move locales to locales/ directory
