+++ src/app/locales/es.ts
@@ -1,94 +1,102 @@
-import type Localization from "./localization";
+import type Localization from "../i18n/localization";
 
 export default {
   common: {
     app_name: "Predictive Text Studio",
-    back: "Back",
-    help: "Help",
+    back: "Atrás",
+    help: "Ayuda",
     excel: "Excel",
     google_sheets: "Google Sheets",
     tsv: "TSV",
-    example: "Example",
-    upload: "Upload",
+    example: "Ejemplo",
+    upload: "Carga",
     direct_entry: "Direct entry",
     error: "Error",
-    bcp47_tag: "BCP47 Tag",
-    name: "Name",
-    size: "Size",
-    type: "Type",
-    actions: "Actions",
-    table_name: "Table Name",
-    word: "Word",
-    count: "Count",
-    optional: "Optional",
-    save: "Save",
-    close: "Close",
-    or: "or",
+    bcp47_tag: "Código de idioma BCP-47",
+    name: "Nombre",
+    size: "Tamaño",
+    type: "Tipo",
+    actions: "Accions",
+    table_name: "Nombre de tabla",
+    word: "Palabra",
+    count: "Ocurencias",
+    optional: "Opcionál",
+    save: "Guardar",
+    close: "Cerrar",
+    or: "o",
   },
   input: {
     enter_bcp47_tag: "Enter the BCP 47 Tag",
-    invalid_bcp47_tag: "Invalid BCP 47 Tag",
-    download_kmp_package: "Download KMP Package",
-    read_values: "Read Values",
-    connection_error: "Could not connect to Google Sheets",
+    invalid_bcp47_tag: "Código de idioma no valido",
+    download_kmp_package: "Descara archivo KMP",
+    read_values: "Lee valors",
+    connection_error: "No se pudo conectar a Hojas de Cálculo de Google",
     invalid_google_sheets_url:
-      "is not a valid Google Sheets URl. Please paste a valid one.",
-    language: "Language",
-    author_or_organization: "Author or Organization",
-    shortcode: "Shortcode",
-    dictionary_name: "Dictionary Name",
-    model_id: "Model ID",
-    copyright: "Copyright",
-    keyboard_preview: "Keyboard Preview",
-    keyboard_preview_alt: "An iOS keyboard in English",
-    add_source: "Add Source",
-    add_row: "Add Row",
-    drag_and_drop: "Drag and drop an Excel .xlsx or TSV file here!",
-    browse_file: "Browse file",
-    advanced_options: "Advanced options",
-    only_works_with: "*Only works with .xlsx files and Google Sheets",
-    spreadsheet_word_column: "Spreadsheet Word Column",
-    spreadsheet_count_column: "Spreadsheet Count Column",
-    no_custom_word_column: "No custom word column specified! Defaulting to",
-    no_custom_count_column: "No custom count column specified! Defaulting to",
+      "no es un URL valído de Hojas de Cálculo de Google. Por favor, trata de nuevo.",
+    language: "Idioma",
+    author_or_organization: "Autor u organización",
+    shortcode: "Código corto",
+    dictionary_name: "Nombre de diccionario",
+    model_id: "Código de modelo",
+    copyright: "Derechos",
+    keyboard_preview: "Vista previa",
+    keyboard_preview_alt: "Un teclado iOS para inglés",
+    add_source: "Agrega diccionario",
+    add_row: "Agrega fila",
+    drag_and_drop: "¡Arrasta un archivo Excel .xlsx o TSV file aquí!",
+    browse_file: "carga archivo",
+    advanced_options: "Opciones avacadas",
+    only_works_with:
+      "*Sólo funcciona con archivos Excel .xlsx y Hojas de Cálculo de Google",
+    spreadsheet_word_column: "la columna que contiene palabras",
+    spreadsheet_count_column: "la columna que contiene occurencias",
+    no_custom_word_column:
+      "¡No has escojído la columna que contiene las palabras! Usando columna",
+    no_custom_count_column:
+      "¡No has escojído la columna que contiene las occurencias! Usando columna",
   },
   page: {
     help: {
-      title: "What Do I Upload?",
+      title: "¿Qué tengo que cargar?",
       intro:
-        'Predictive Text Studio uses the spreadsheet you provide to customize the words in your keyboard\'s dictionary. The "Word" column contains frequently used words in your language that you would like to add to your dictionary. The "Count" column gives the amount of times or frequency that a word appears in your text.',
+        "Predictive Text Studio usa la hoja de cálculo que cargas en el diccionario de tu teclado. " +
+        "La columna “palabras” contiene palabras de tu idioma cuales quieres sugirir en tu tecaldo predictivo. " +
+        "La columna “occurrencias” opcionalmente da las veces que ocurre la palabra en tu idioma.",
       excel_instruction:
         'Create an Excel file with "Word" in the A column and "Count" in the B column. You may upload the Excel file directly to the Upload Area.',
       google_sheets_instruction:
         'Create a Google Sheets spreadsheet with "Word" in the A column and "Count" in the B column. When you have completed the spreadsheet press the "Share" button in the top right corner of the Google Sheets interface and copy a link which "Anyone on the internet with this link can view". You may now paste this link in the Google Sheets URL input.',
       tsv_instruction:
-        'TSV stands for Tab Separated Values. Like Excel or Google Sheets, create a text file with "Word" in the first column and "Count" in the second column. Make sure your file has a .tsv extension and upload it to the File Uploader. For exhaustive details on how your file should look, please visit the',
-      keyman: "Keyman Website",
+        "TSV (en inglés: Tab Separated Values) es un archivo que contiene valores separados por tabulaciones. " +
+        "Así como Excel o Hojas de Cálculo de Google, " +
+        'create a text file with "Word" in the first column and "Count" in the second column. Segurate que tu archivo tiene la extencion “.tsv” y cargala en la area de cargas. ' +
+        "Para detalles técnicas sobre el formato .TSV, por favor hecha un vistazo al",
+      keyman: "sito web de Keyman",
     },
     main: {
-      title: "Welcome to Predictive Text Studio",
-      designed_for: "Designed for",
-      add: "Add",
-      prediction: "prediction",
-      and: "and",
-      autocorrect: "autocorrect",
-      to_your_language: "to your language",
-      get_started: "Get started",
-      step_one: "Step 1: Enter your language",
-      step_two: "Step 2: Add a word list",
-      customize: "Customize",
-      privacy_policy: "Privacy Policy",
-      about_the_team: "About the Team",
+      title: "Le damos la bienvenida a Predictive Text Studio",
+      designed_for: "diseñado para",
+      add: "Agrega",
+      prediction: "sugerencias",
+      and: "y",
+      autocorrect: "autocorrección",
+      to_your_language: "para tu idioma",
+      get_started: "Empezamos",
+      step_one: "Primero: elige tu idioma",
+      step_two: "Segundo: Agrega una lista de palabras",
+      customize: "Configurar",
+      privacy_policy: "Política de privacidad",
+      about_the_team: "Sobre el equipo",
     },
     lang: {
-      go_back_to_main_page: "Go back to main page",
-      information: "Information",
-      sources: "Sources",
-      download: "Download",
+      go_back_to_main_page: "Regresa a la página principal",
+      information: "Información",
+      sources: "Fuentes de diccionario",
+      download: "Descarga",
     },
     privacy: {
-      title: "Predictive Text Studio Privacy Policy",
-      effective_as_of: "Effective as of",
+      title: "Política de Privacidada de Predictive Text Studio",
+      effective_as_of: "A partir de",
       introduction: {
         title: "Introduction",
         head: {
@@ -169,10 +177,10 @@ export default {
       },
     },
     team: {
-      title: "Team",
-      intro: "These are people who have contributed to the project",
-      core_maintainers: "Core Maintainers",
-      team_members: "Team Members",
+      title: "Equipo",
+      intro: "La gente que han desarrollado Predictive Text Studio",
+      core_maintainers: "Mantenador principal",
+      team_members: "Miembros de equipo",
     },
   },
 } as Localization;
